Moved Post Getting and Validation to Separate Files
