+++ final-project/app.ts
@@ -1,9 +1,10 @@
 import pug from "pug";
 import express from "express";
-import { AtpAgent } from "@atproto/api";
 import Post from "./post";
 import cookieParser from "cookie-parser";
 import Credentials from "./credentials";
+import { getAgent, getLogin } from "./agent";
+import { getPosts, getSearchTerms } from "./postGetters";
 
 const app = express();
 const port = process.env.PORT || 5001;
@@ -49,18 +50,22 @@ app.get("/whats-hot", async (req, res) => {
 
 app.post("/search", async (req, res) => {
   const term: string = req.body.term;
-  let searchTerms: string[] = getSearchTerms(req);
-  if (searchTerms.includes(term)) {
-    searchTerms = searchTerms.filter((e) => e != term);
+  if (term.length == 0) {
+    res.redirect(302, "/whats-hot");
+  } else {
+    let searchTerms: string[] = getSearchTerms(req);
+    if (searchTerms.includes(term)) {
+      searchTerms = searchTerms.filter((e) => e != term);
+    }
+    searchTerms.reverse().push(term);
+    searchTerms.reverse();
+    if (searchTerms.length > 5) {
+      searchTerms.pop();
+    }
+    res
+      .cookie("searchTerms", JSON.stringify(searchTerms))
+      .redirect(302, "/topic");
   }
-  searchTerms.reverse().push(term);
-  searchTerms.reverse();
-  if (searchTerms.length > 5) {
-    searchTerms.pop();
-  }
-  res
-    .cookie("searchTerms", JSON.stringify(searchTerms))
-    .redirect(302, "/topic");
 });
 
 app.get("/topic", async (req, res) => {
@@ -86,14 +91,16 @@ app.get("/topic", async (req, res) => {
 });
 
 app.post("/like", async (req, res) => {
-  const agent = await getAgent(req);
+  const credentials = getLogin(req);
+  const agent = await getAgent(credentials);
   const postInfo = JSON.parse(req.body.post);
   await agent.like(postInfo.uri, postInfo.cid);
   res.redirect(302, "/whats-hot");
 });
 
 app.post("/comment", async (req, res) => {
-  const agent = await getAgent(req);
+  const credentials = getLogin(req);
+  const agent = await getAgent(credentials);
   // Add code to comment here
   console.log(req.body.comment);
   res.redirect(302, "/whats-hot");
@@ -106,57 +113,3 @@ app.use((req, res) => {
 app.listen(port, () => {
   console.log("Server running at http:localhost:" + port);
 });
-
-function getLogin(req: any): Credentials {
-  const result: Credentials = JSON.parse(req.cookies.credentials);
-  return result;
-}
-
-async function getAgent(req): Promise<AtpAgent> {
-  const agent = new AtpAgent({
-    service: "https://bsky.social",
-  });
-  await agent.login(getLogin(req));
-  return agent;
-}
-
-async function getPosts(
-  req,
-  searchTerm: string | undefined = undefined
-): Promise<Post[] | undefined> {
-  let result: Post[] | undefined = undefined;
-  try {
-    const agent = await getAgent(req);
-    result =
-      searchTerm == undefined
-        ? await getHotPosts(agent)
-        : await getTopicalPosts(agent, searchTerm);
-  } catch (e) {}
-  return result;
-}
-
-async function getHotPosts(agent: AtpAgent): Promise<Post[]> {
-  const feed = await agent.app.bsky.feed.getFeed({
-    feed: "at://did:plc:z72i7hdynmk6r22z27h6tvur/app.bsky.feed.generator/whats-hot",
-    limit: 100,
-  });
-  return feed.data.feed.map((e) => new Post(agent, e.post));
-}
-
-async function getTopicalPosts(agent: AtpAgent, term: string): Promise<Post[]> {
-  const feed = await agent.app.bsky.feed.searchPosts({
-    q: term,
-    limit: 100,
-  });
-  return feed.data.posts.map((e) => new Post(agent, e));
-}
-
-function getSearchTerms(req) {
-  let result: string[] = [];
-  try {
-    result = JSON.parse(req.cookies.searchTerms);
-  } catch (e) {
-    result = [];
-  }
-  return result;
-}
